pyproject.toml
[project]
name = "simple-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.18.0",
    "langchain-community>=0.3.24",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.5",
    "pinecone>=7.0.2",
    "pypdf>=5.5.0",
    "python-dotenv>=1.1.0"
]

[tool.setuptools.packages.find]
exclude = ["data*"]